Added 'params' field
The 'params' field was added in both CircuitDefinition and
CompoenntDefinition's 'circuit' field. It specifies a map of component
IDs and a map of parameter name and value (which is a JSON Any type).
They get mapped (if there are any params) to all prebuilt components
(and only prebuilt components) during component instantiation.