Added initial_evaluate()
All components are now iterated through and evaluated in init(). This is
to mitigate an 'initial propagation' bug that makes a defaulted NAND not
propagate its 'on' output at simulation start. The reason we did not add
NAND to be a source component is because we might later allow custom
initial states for components passed through by the Params options. This
is the cleanest and future-proof solution.